+++ IoTPublishShare/src/com/aalto/protocol/design/iotps/objects/IoTPSAckObject.java
@@ -2,22 +2,22 @@ package com.aalto.protocol.design.iotps.objects;
 
 public class IoTPSAckObject {
 
-	private int sub_seq_no;
-	private int seq_no;
+	private double sub_seq_no;
+	private double seq_no;
 	
 	private String fromIp;
 	private int fromPort;
 	
-	public int getSubSeqNo() {
+	public double getSubSeqNo() {
 		return sub_seq_no;
 	}
-	public void setSubSeqNo(int sub_seq_no) {
+	public void setSubSeqNo(double sub_seq_no) {
 		this.sub_seq_no = sub_seq_no;
 	}
-	public int getSeqNo() {
+	public double getSeqNo() {
 		return seq_no;
 	}
-	public void setSeqNo(int seq_no) {
+	public void setSeqNo(double seq_no) {
 		this.seq_no = seq_no;
 	}
 	public String getFromIp() {
