+++ IoTPublishShare/src/com/aalto/protocol/design/iotps/objects/IoTPSUpdateObject.java
@@ -6,8 +6,8 @@ public class IoTPSUpdateObject extends IoTPSObject{
 	private String client_ip;
 	private int client_port;
 	private String device_id;
-	private int seq_no;
-	private int sub_seq_no;
+	private double seq_no;
+	private double sub_seq_no;
 	private double timestamp;
 	private String sensor_data;
 	private int version;
@@ -22,11 +22,11 @@ public class IoTPSUpdateObject extends IoTPSObject{
 	public String getDeviceId() {return device_id;}
 	public void setDeviceId(String id) {this.device_id = id;}
 	
-	public int getSeqNo() {return seq_no;}
-	public void setSeqNo(int seq) {this.seq_no = seq;}
+	public double getSeqNo() {return seq_no;}
+	public void setSeqNo(double seq) {this.seq_no = seq;}
 	
-	public int getSubSeqNo() {return sub_seq_no;}
-	public void setSubSeqNo(int sub) {this.sub_seq_no = sub;}
+	public double getSubSeqNo() {return sub_seq_no;}
+	public void setSubSeqNo(double sub) {this.sub_seq_no = sub;}
 	
 	public double getTimestamp() {return timestamp;}
 	public void setTimestamp(double ts) {this.timestamp = ts;}
